PR feedback

+++ concordium-node/src/p2p/maintenance.rs
@@ -373,7 +373,7 @@ impl P2PNode {
             events_queue_size: conf.connection.events_queue_size,
             deduplication_hashing_algorithm: conf.connection.deduplication_hashing_algorithm,
             regenesis_arc,
-            max_normal_keep_alive_ms: (conf.connection.max_normal_keep_alive as u64) * 1000,
+            max_normal_keep_alive_ms: conf.connection.max_normal_keep_alive * 1000,
         };
 
         let connection_handler = ConnectionHandler::new(conf, server);
